make ci work for branch, too
git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@392479 13f79535-47bb-0310-9956-ffa450edef68

+++ ci.sh
@@ -179,7 +179,11 @@ fi
 
     mkdir -p $DEPLOY_DIR > /dev/null 2>&1
 
-    cp $DIR/maven-components/maven-cli/target/*.tar.gz $DEPLOY_DIR/$DIST
+    if [ -f $DIR/maven-components/maven-cli ]; then
+      cp $DIR/maven-components/maven-cli/target/*.tar.gz $DEPLOY_DIR/$DIST
+    else
+      cp $DIR/maven-components/maven-core/target/*.tar.gz $DEPLOY_DIR/$DIST
+    fi
 
   else
